# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Screen recorder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * private = token rotated on switch, so old links die; unlisted = playable with the share token; public = playable by anyone who knows the id.
 */
export type Visibility = typeof Visibility[keyof typeof Visibility];


export const Visibility = {
  private: 'private',
  unlisted: 'unlisted',
  public: 'public',
} as const;

export interface Video {
  id: string;
  title: string;
  filename: string;
  mimeType: string;
  /** File size in bytes */
  size: number;
  /**
     * Duration in seconds
     * @nullable
     */
  duration?: number | null;
  createdAt: string;
  visibility: Visibility;
  /** Required as ?t= to stream anything that is not public */
  shareToken: string;
  /**
     * Path inside the configured Dropbox folder, once mirrored
     * @nullable
     */
  dropboxPath?: string | null;
  /**
     * Dropbox shared link, present while visibility allows one
     * @nullable
     */
  dropboxSharedUrl?: string | null;
  /** @nullable */
  dropboxUploadedAt?: string | null;
}

export interface VideoUpload {
  title?: string;
}

export interface VideoUpdate {
  /** @minLength 1 */
  title?: string;
  visibility?: Visibility;
}

export interface DropboxStatus {
  /** Whether Dropbox credentials are present in the environment */
  configured: boolean;
  /** Whether those credentials actually authenticate */
  connected: boolean;
  folder: string;
  /** @nullable */
  accountEmail?: string | null;
  /** @nullable */
  error?: string | null;
}

export interface DropboxFile {
  id: string;
  name: string;
  path: string;
  size: number;
  /** @nullable */
  clientModified?: string | null;
  /** @nullable */
  sharedUrl?: string | null;
}

export interface DropboxListing {
  files: DropboxFile[];
  /** @nullable */
  error?: string | null;
}

export interface VideoStats {
  totalCount: number;
  totalSizeBytes: number;
  totalDurationSeconds: number;
}

export type GetVideoParams = {
/**
 * Share token; required for recordings that are not public
 */
t?: string;
};