# lib/api-zod/src/generated/types/dropboxFile.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Screen recorder API
 * OpenAPI spec version: 0.1.0
 */

export interface DropboxFile {
  id: string;
  name: string;
  path: string;
  size: number;
  /** @nullable */
  clientModified?: Date | null;
  /** @nullable */
  sharedUrl?: string | null;
}