# lib/api-zod/src/generated/types/dropboxListing.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Screen recorder API
 * OpenAPI spec version: 0.1.0
 */
import type { DropboxFile } from './dropboxFile';

export interface DropboxListing {
  files: DropboxFile[];
  /** @nullable */
  error?: string | null;
}