# lib/api-zod/src/generated/types/dropboxStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Screen recorder API
 * OpenAPI spec version: 0.1.0
 */

export interface DropboxStatus {
  /** Whether Dropbox credentials are present in the environment */
  configured: boolean;
  /** Whether those credentials actually authenticate */
  connected: boolean;
  folder: string;
  /** @nullable */
  accountEmail?: string | null;
  /** @nullable */
  error?: string | null;
}