# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Screen recorder API
 * OpenAPI spec version: 0.1.0
 */
import type { Visibility } from './visibility';

export interface Video {
  id: string;
  title: string;
  filename: string;
  mimeType: string;
  /** File size in bytes */
  size: number;
  /**
     * Duration in seconds
     * @nullable
     */
  duration?: number | null;
  createdAt: Date;
  visibility: Visibility;
  /** Required as ?t= to stream anything that is not public */
  shareToken: string;
  /**
     * Path inside the configured Dropbox folder, once mirrored
     * @nullable
     */
  dropboxPath?: string | null;
  /**
     * Dropbox shared link, present while visibility allows one
     * @nullable
     */
  dropboxSharedUrl?: string | null;
  /** @nullable */
  dropboxUploadedAt?: Date | null;
}