# lib/api-zod/src/generated/types/videoStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Screen recorder API
 * OpenAPI spec version: 0.1.0
 */

export interface VideoStats {
  totalCount: number;
  totalSizeBytes: number;
  totalDurationSeconds: number;
}