# lib/api-zod/src/generated/types/videoUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Screen recorder API
 * OpenAPI spec version: 0.1.0
 */
import type { Visibility } from './visibility';

export interface VideoUpdate {
  /** @minLength 1 */
  title?: string;
  visibility?: Visibility;
}