# lib/api-zod/src/generated/types/visibility.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Screen recorder API
 * OpenAPI spec version: 0.1.0
 */

/**
 * private = token rotated on switch, so old links die; unlisted = playable with the share token; public = playable by anyone who knows the id.
 */
export type Visibility = typeof Visibility[keyof typeof Visibility];


export const Visibility = {
  private: 'private',
  unlisted: 'unlisted',
  public: 'public',
} as const;